new features - tool to make your own palette - add function deprecation - add attributed to cmap to more color format
