+++ .gitignore
@@ -1,4 +1,4 @@
-venv
+venv/
 __pycache__
 dist
 build
@@ -6,3 +6,4 @@ pypalettes.egg-info
 site
 .pytest_cache
 sandbox.ipynb
+.vscode/
