--- .vscode/settings.json
@@ -1,7 +1,0 @@
-{
-   "python.testing.pytestArgs": [
-      "tests"
-   ],
-   "python.testing.unittestEnabled": false,
-   "python.testing.pytestEnabled": true
-}
