+++ pypalettes/__init__.py
@@ -1,4 +1,6 @@
 
-from .main import load_cmap, get_source, get_hex, get_rgb, get_kind
+from .main import load_cmap
+from .deprecated import get_source, get_hex, get_rgb, get_kind
+from .make import make_cmap
 
-__all__ = ['load_cmap, get_source, get_hex, get_rgb, get_kind']
+__all__ = ['load_cmap', 'make_cmap']
