+++ pyproject.toml
@@ -4,7 +4,7 @@ version = "0.1.3"
 authors = [
   { name="Joseph Barbier", email="joseph.barbierdarnal@gmail.com" },
 ]
-description = "A large collection of color maps for Python"
+description = "A large (+2500) collection of color maps for Python"
 readme = "README.md"
 requires-python = ">=3.9"
 classifiers = [
