--- requirements-dev.txt
@@ -1,3 +1,0 @@
-matplotlib==3.9.0
-pytest==8.2.2
-bs4==0.0.2
