+++ requirements.py
@@ -9,26 +9,14 @@
 import subprocess
 
 packages = [
-    "matplotlib"
-]
-packages_dev = packages + \
-[
+    "matplotlib",
     "pytest",
     "bs4"
 ]
 REQUIREMENTS_FILE = "requirements.txt"
-REQUIREMENTS_FILE_DEV = "requirements-dev.txt"
 
 
 def get_package_version(package_name: str) -> str:
-    """
-    Get the version of a package using pip show.
-    Args:
-        package_name: Name of the package to get the version for.
-    Returns:
-        The version of the package.
-    """
-
     result = subprocess.run(
         ["pip", "show", package_name], capture_output=True, text=True
     )
@@ -49,15 +37,4 @@ with open(REQUIREMENTS_FILE, "w") as file:
         else:
             print(f"Version not found for package {package}")
     print(f"{REQUIREMENTS_FILE} file updated\n\n")
-
-with open(REQUIREMENTS_FILE_DEV, "w") as file:
-    for package in packages_dev:
-        version = get_package_version(package)
-        if version:
-            packageVersion = f"{package}=={version}"
-            print(packageVersion)
-            file.write(packageVersion + "\n")
-        else:
-            print(f"Version not found for package {package}")
-    print(f"{REQUIREMENTS_FILE_DEV} file updated")
