+++ requirements.txt
@@ -1,1 +1,3 @@
 matplotlib==3.9.0
+pytest==8.2.2
+bs4==0.0.2
